# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Cast API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/notifications/unread": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get unread notifications",
                "operationId": "22cc6d096c61ed4ca041d76935f38ba1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/v1/show/{id}": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Get show info",
                "operationId": "ee8d60d21b7a920409c7fd8084bdfce8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/v1/show/short": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Get user short shows list",
                "operationId": "04d54ff4e9a8aa908791ade2977e2780",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/v1/show/list": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Get user shows",
                "operationId": "44062dd4c0e5c48066113519ecc46e23",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/v1/show/create": {
            "put": {
                "tags": [
                    "Shows"
                ],
                "summary": "Update show",
                "operationId": "7a2e3f754bdab81470cb92483aa4b3c4",
                "requestBody": {
                    "description": "Update show",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "format",
                                    "timezone",
                                    "language",
                                    "explicit",
                                    "category",
                                    "show_id"
                                ],
                                "properties": {
                                    "show_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Exapmle show name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Exapmle show description"
                                    },
                                    "artwork": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "format": {
                                        "type": "string",
                                        "example": "episodic"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Etc/GMT"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "explicit": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "sport"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Author name"
                                    },
                                    "podcast_owner": {
                                        "type": "string",
                                        "example": "example author"
                                    },
                                    "email_owner": {
                                        "type": "string",
                                        "example": "example@domain.com"
                                    },
                                    "copyright": {
                                        "type": "string",
                                        "example": "copyright"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shows"
                ],
                "summary": "Create new show",
                "operationId": "af5c356e6f0d6aee9bd8fc7409a8ec44",
                "requestBody": {
                    "description": "Create show",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "format",
                                    "timezone",
                                    "language",
                                    "explicit",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Exapmle show name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Exapmle show description"
                                    },
                                    "artwork": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "format": {
                                        "type": "string",
                                        "example": "episodic"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Etc/GMT"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "explicit": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "sport"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Author name"
                                    },
                                    "podcast_owner": {
                                        "type": "string",
                                        "example": "example author"
                                    },
                                    "email_owner": {
                                        "type": "string",
                                        "example": "example@domain.com"
                                    },
                                    "copyright": {
                                        "type": "string",
                                        "example": "copyright"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users list",
                "operationId": "703f8f4a96c6970997b1f19ad624bd81",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "User detail",
                "operationId": "93cf0b63e1b9467a1edca8a3cd5fcf52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        }
    }
}